edition enums
