+++ backend/src/feature-flags/setTenantProperties.ts
@@ -1,12 +1,13 @@
 import { PostHog } from 'posthog-node'
 import { API_CONFIG, POSTHOG_CONFIG } from '../config'
+import { Edition } from '../types/common'
 
 export default async function setPosthogTenantProperties(
   tenant: any,
   posthog: PostHog,
   database: any,
 ) {
-  if (POSTHOG_CONFIG.apiKey && API_CONFIG.edition === 'crowd-hosted') {
+  if (POSTHOG_CONFIG.apiKey && API_CONFIG.edition === Edition.CROWD_HOSTED) {
     const automationCount = await database.automation.count({
       where: {
         tenantId: tenant.id,
