+++ backend/src/segment/identifyTenant.ts
@@ -1,12 +1,13 @@
 import { SEGMENT_CONFIG, API_CONFIG } from '../config'
 import setPosthogTenantProperties from '../feature-flags/setTenantProperties'
+import { Edition } from '../types/common'
 
 export default async function identifyTenant(req) {
   if (SEGMENT_CONFIG.writeKey) {
     const Analytics = require('analytics-node')
     const analytics = new Analytics(SEGMENT_CONFIG.writeKey)
 
-    if (API_CONFIG.edition === 'crowd-hosted') {
+    if (API_CONFIG.edition === Edition.CROWD_HOSTED) {
       analytics.group({
         userId: req.currentUser.id,
         groupId: req.currentTenant.id,
@@ -14,7 +15,7 @@ export default async function identifyTenant(req) {
           name: req.currentTenant.name,
         },
       })
-    } else if (API_CONFIG.edition === 'community') {
+    } else if (API_CONFIG.edition === Edition.COMMUNITY) {
       if (!req.currentUser.email.includes('crowd.dev')) {
         analytics.group({
           userId: req.currentUser.id,
