--- backend/src/types/featureFlag.ts
@@ -1,6 +1,0 @@
-export enum FeatureFlag {
-  AUTOMATIONS = 'automations',
-  COMMUNITY_HELP_CENTER_PRO = 'community-help-center-pro',
-  EAGLE_EYE = 'eagle-eye',
-  ORGANIZATIONS = 'organizations',
-}
